fix(http3): make the WebTransport client actually compile on dev
CWIST_WEBTRANSPORT was never defined by any build, so every #ifdef
CWIST_WEBTRANSPORT region in http3.c and http3_client.c had gone
uncompiled since it was written. Turning it on surfaced three real breaks
against the lsquic WebTransport branch head (litespeedtech/lsquic#629,
5b7dc8f8), which dev now pins instead of the stale 06fffaf8 merge:

- wti_on_session_close takes uint32_t for the application error code, not
  uint64_t (http3.c, http3_client.c). WebTransport error codes are 32-bit
  on the wire; the branch narrowed the callback to match.
- wti_on_stream_ss_code returns uint32_t, not uint64_t (http3.c).
- es_max_webtransport_sessions was 4 in the client. lsquic_engine.c
  rejects anything above 1, so engine creation failed outright with
  "only supports 1 session per connection". The server already set 1.

The Makefile now defines CWIST_WEBTRANSPORT by default on this branch
(CWIST_WEBTRANSPORT=0 opts out), so the code stays compiled from here on.

Also: objects now depend on a .build-flags stamp holding CFLAGS. Flipping
CWIST_WEBTRANSPORT changes struct members in http3.c, and make had no
reason to rebuild on a flag change alone, so a flipped build produced a
library with objects compiled both ways. The mismatched layouts did not
fail to link, they smashed the stack at runtime. A changed flag now forces
a full rebuild.

Verified: clean build with WebTransport on, test_http3 passes 12 runs out
of 12 (the smash was intermittent, so one pass proves nothing), 10 WT
client symbols present in libcwist.a, raw allocator gate passes.

+++ src/net/http/http3.c
@@ -1817,9 +1817,13 @@ static void cwist_h3_wt_on_session_rejected(void *ctx,
     (void)reason_len;
 }
 
+/* WebTransport application error codes are 32-bit on the wire. The lsquic
+ * WebTransport branch declared these two callbacks with uint64_t when this
+ * code was written; the branch head (litespeedtech/lsquic#629) narrows both
+ * to uint32_t. */
 static void cwist_h3_wt_on_session_close(lsquic_wt_session_t *sess,
                                          lsquic_wt_session_ctx_t *sess_ctx,
-                                         uint64_t code, const char *reason,
+                                         uint32_t code, const char *reason,
                                          size_t reason_len) {
     (void)sess;
     (void)code;
@@ -1883,7 +1887,7 @@ static void cwist_h3_wt_on_stream_close(lsquic_stream_t *stream,
     cwist_wt_handle_free((cwist_wt_handle_t *)st_h);
 }
 
-static uint64_t cwist_h3_wt_on_stream_ss_code(lsquic_stream_t *stream,
+static uint32_t cwist_h3_wt_on_stream_ss_code(lsquic_stream_t *stream,
                                               lsquic_stream_ctx_t *st_h) {
     (void)stream;
     (void)st_h;
